review: extract cueProfileSet helper, clean up cancel test
Three Section H tests duplicated the buy→charge→collect→cue setup
verbatim. The new cancel test additionally dropped the collect_id
error guard from the copy, leaving a silent failure path.

- Move the shared setup into cueProfileSet() in _helpers.ts (with the
  guard), import and use it across all three Section H tests
- Drop the vacuous selectedBefore assertion (the test itself sets the
  class; asserting it's there tests the harness, not production code)
- Replace the selectedAfter page.evaluate with a Playwright locator so
  the assertion auto-retries and costs one fewer CDP round-trip
- Replace the implementation-trace comment in the test with the
  observable contract
- Trim the source comment in Database.ts to the UX invariant
